requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
gunicorn
python-dotenv
Flask-Migrate

